Use Page for all & consolidate tmpls
