+++ views/templates/mainLogin.tmpl
@@ -1,7 +1,7 @@
 
 {{define "content"}}
 
-{{template "header"}}
+{{template "header" .}}
 				
 			</div>
